This is the illustration of simulation 2 results.
It is visually verified that our method works well. Comparing the result with the decomposition obtained from solving RPCA, we see that RPCA only identifies the moving hotspot, but fails to separate the true background and the static hotspot. This is because the RPCA does not consider the spatial smoothness of the background.