This is the illustration of case study result.
Using SSD method, some Calcium accumulations with low intensity are not as anomalie, and the edges of the Calcium accumulations are rough across consecutive images, because SSD does not take the similarity of the Calcium accumulations’ locations into consideration. This case study further illustrates the versatility of the ATD framework and demonstrates its ability to solving real-world problems.