This is the illustration of simulation 1 results.
We see that the SSD only captures some points on the crack because the smoothness of the crack in the temporal mode is ignored. The l_1 regularization in the SSD encourages sparsity of the anomaly but pushes all pixels with low intensity to zero, including the ones on the crack. In contrast, the ATD-based method regularizes the temporal smoothness of X_2, which preserves the pixels with low intensity on the crack, whereas regularize other pixels to zeros.