This is the illustration of case study results.
The result shows that the smooth change of AVC regions across different images is captured by the proposed ATD method, which reflects the anatomic reality. For comparison, we also conduct the SSD to extract the anomaly regions on individual images. Some AVC regions with low intensity are not identified as anomalies (see the red circles), and the edges of the AVC regions are rough across consecutive images because the SSD does not take the similarity of the AVC regions’ locations into consideration. This case study further illustrates the versatility of the ATD framework and demonstrates its ability to solving real-world problems.